[FEATURE] Example to: And I want that the "HTML Body" allows HTML tags.
We are in the TYPO3 context. Therefore you should orientate yourself on the form framework and to be more precise on the confirmation finisher of the form extension. Here, a content element (e.g. text) is simply used for the HTML part and the corresponding UID is saved here as content in the form. This can then be used in the finisher to pass the HTML markup to the template.